Apply comments yet again (#47)
