+++ src/pipeline/evaluation/structure_metrics.py
@@ -409,9 +409,10 @@ def calculate_structure_similarity_metrics(
     if device.type == "cuda" and not torch.cuda.is_available():
         raise RuntimeError("Structure metrics requested CUDA, but torch.cuda.is_available() is False")
 
-    print(f"Structure metric tensors use {device}; SimpleITK displacement registration and mask resampling run on CPU.")
+    #To add logging if needed
+    #print(f"Structure metric tensors use {device}; SimpleITK displacement registration and mask resampling run on CPU.")
 
-    for patient_id, gen_paths in tqdm(generated.items(), desc="Calculating generated vs real structure metrics."):
+    for patient_id, gen_paths in tqdm(generated.items(), desc="Calculating generated vs real structure metrics"):
         all_ref_paths = originals.get(patient_id, [])
         ref_paths = [path for path in all_ref_paths if not _is_planning_ct_path(path)]
         planning_paths = [path for path in all_ref_paths if _is_planning_ct_path(path)]
